update : 型情報をまとめた
