+++ src/features/courses/types/CardCarousel.tsx
@@ -1,0 +1,11 @@
+export type Course = {
+  id: number;
+  title: string;
+  description: string;
+  achievementLevel: number;
+  maxAchievementLevel: number;
+};
+
+export type CardCarouselProps = {
+  courses: Course[];
+};
